+++ test/external/external_test_opt.py
@@ -4,7 +4,6 @@ import numpy as np
 import torch
 
 from tinygrad import nn, GlobalCounters, Tensor, Device
-from tinygrad.nn import optim
 from tinygrad.nn.state import get_parameters
 from tinygrad.engine.realize import capturing
 
@@ -120,27 +119,6 @@ class TestOpt(unittest.TestCase):
       d.realize()
     np.testing.assert_allclose(d.numpy(), na*nb+nc, rtol=1e-5, atol=1e-7)
 
-  def test_fold_batchnorm(self):
-    with Tensor.train():
-      img = Tensor.ones(1,32,4,4).contiguous()
-      bn = nn.BatchNorm2d(32, track_running_stats=False)
-      with CLCache() as cache:
-        img_bn = bn(img).realize()
-        print(img_bn)
-        assert cache.count == 3, f"optimizer didn't fold batchnorm, got {cache.count}"
-
-  def test_fold_conv_batchnorm_sgd(self):
-    with Tensor.train():
-      img = Tensor.ones(1,3,4,4)
-      c1 = nn.Conv2d(3,32,3)
-      bn = nn.BatchNorm2d(32, track_running_stats=False)
-      opt = optim.SGD(get_parameters([c1, bn]))
-      with CLCache(allowed=16): # this is too high
-        img_bn = bn(c1(img)).elu().sum()
-        opt.zero_grad()
-        img_bn.backward()
-        opt.step()
-
   def test_fold_conv_batchnorm_notrain(self):
     img = Tensor.ones(1,3,8,8)
     c1 = nn.Conv2d(3,32,3)
@@ -150,16 +128,6 @@ class TestOpt(unittest.TestCase):
     with CLCache() as cache:
       bn(c1(img)).relu().realize()
       assert cache.count == 1, f"optimizer didn't fold conv-batchnorm at test time, got {cache.count}"
-
-  def test_fold_conv_batchnorm(self):
-    with Tensor.train():
-      img = Tensor.ones(1,3,8,8)
-      c1 = nn.Conv2d(3,32,3)
-      bn = nn.BatchNorm2d(32, track_running_stats=False)
-      with CLCache() as cache:
-        img_conv = bn(c1(img)).relu().realize()
-        print(img_conv)
-        assert cache.count == 4, f"optimizer didn't fold conv-batchnorm, got {cache.count}"
 
   def test_fold_conv_elu(self):
     img = Tensor.ones(1,4,8,8)
@@ -174,15 +142,6 @@ class TestOpt(unittest.TestCase):
     img = Tensor.ones(1,4,8,8)
     c1 = nn.Conv2d(4, 4, kernel_size=3)
     c2 = nn.Conv2d(4, 4, kernel_size=3)
-    with CLCache() as cache:
-      img_conv = img.sequential([c1, Tensor.relu, c2, Tensor.relu]).realize()
-      print(img_conv)
-      assert cache.count == 2, "optimizer didn't fold conv/relu"
-
-  def test_fold_conv_relu_nobias(self):
-    img = Tensor.ones(1,4,8,8)
-    c1 = nn.Conv2d(4, 4, kernel_size=3, bias=False)
-    c2 = nn.Conv2d(4, 4, kernel_size=3, bias=False)
     with CLCache() as cache:
       img_conv = img.sequential([c1, Tensor.relu, c2, Tensor.relu]).realize()
       print(img_conv)
