+++ test/test_schedule.py
@@ -194,6 +194,26 @@ class TestSchedule(unittest.TestCase):
       out = bn(img)
       check_schedule(out, 3)
 
+  def test_fold_conv_batchnorm(self):
+    with Tensor.train():
+      img = Tensor.empty(1,3,8,8)
+      c1 = nn.Conv2d(3,32,3)
+      bn = nn.BatchNorm2d(32, track_running_stats=False)
+      out = bn(c1(img)).relu()
+      check_schedule(out, 4, [c1.weight, c1.bias])
+
+  def test_fold_conv_batchnorm_sgd(self):
+    with Tensor.train():
+      img = Tensor.ones(1,3,4,4)
+      c1 = nn.Conv2d(3,32,3)
+      bn = nn.BatchNorm2d(32, track_running_stats=False)
+      opt = nn.optim.SGD(nn.state.get_parameters([c1, bn]))
+      img_bn = bn(c1(img)).elu().sum()
+      opt.zero_grad()
+      img_bn.backward()
+      # this is too high
+      check_schedule(opt.schedule_step(), 18)
+
   def test_fold_conv_relu(self):
     c1 = nn.Conv2d(3,16,3)
 
@@ -201,6 +221,13 @@ class TestSchedule(unittest.TestCase):
     img = Tensor.ones(2,3,64,64)
     out = c1(img).relu()
     check_schedule(out, 1, [c1.weight, c1.bias])
+
+  def test_fold_conv_relu_nobias(self):
+    img = Tensor.ones(1,4,8,8)
+    c1 = nn.Conv2d(4, 4, kernel_size=3, bias=False)
+    c2 = nn.Conv2d(4, 4, kernel_size=3, bias=False)
+    out = img.sequential([c1, Tensor.relu, c2, Tensor.relu])
+    check_schedule(out, 2, [c1.weight, c2.weight, img])
 
   def test_fold_conv_elu(self):
     c1 = nn.Conv2d(3,16,3)
